target and scheme Confirmation
target is a build configuration, scheme is a project workflow including target and will determine the build process steps, output such as Archives , distribution and test destinations.
Both target and scheme configuration is complex, because specific item like bundle id is needed to be set manually.
So built app can be backup in version folder for dragging to simulators, but adjustment in archive .ipa to drag to iphone devices is not convenient.
Just freeze project folder in external disk folder, and in Xcode relaunch is so far the best choice.